Split slash commands to general file
